refactor: migrate to react router v7, refactor code
