+++ app/components/Details/Details.module.css
