+++ app/context/themeContext.tsx
@@ -1,0 +1,12 @@
+import { createContext, type PropsWithChildren } from 'react';
+import useSetTheme from './useSetTheme';
+
+const ThemeContext = createContext({ theme: 'dark', changeTheme: () => {} });
+
+const ThemeProvider = ({ children }: PropsWithChildren) => {
+  const { theme, changeTheme } = useSetTheme();
+
+  return <ThemeContext.Provider value={{ theme, changeTheme }}>{children}</ThemeContext.Provider>;
+};
+
+export { ThemeContext, ThemeProvider };
