+++ app/context/useSetTheme.ts
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+
+const useSetTheme = () => {
+  const [theme, setTheme] = useState('dark');
+
+  const changeTheme = () => {
+    const newTheme = theme === 'dark' ? 'light' : 'dark';
+    setTheme(newTheme);
+  };
+
+  return { theme, changeTheme };
+};
+
+export default useSetTheme;
