--- src/App.tsx
@@ -1,18 +1,0 @@
-import { Navigate, Route, Routes } from 'react-router';
-import { Main } from './components/Main/Main';
-import NotFound from './components/NotFound/NotFound';
-import { Details } from './components/Details/Details';
-
-const App = () => {
-  return (
-    <Routes>
-      <Route path="/" element={<Navigate to={`/page/1`} />} />
-      <Route path="page/:pageId" element={<Main />}>
-        <Route path="species/:id" element={<Details />} />
-      </Route>
-      <Route path="*" element={<NotFound />} />
-    </Routes>
-  );
-};
-
-export { App };
