--- src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,32 +1,0 @@
-import React, { ReactNode } from 'react';
-
-interface ErrorBoundaryProps {
-  children: ReactNode;
-}
-
-interface ErrorBoundaryState {
-  errorMessage: string;
-}
-
-export default class ErrorBoundary extends React.Component<ErrorBoundaryProps> {
-  state: {
-    errorMessage: string;
-  } = {
-    errorMessage: '',
-  };
-
-  static getDerivedStateFromError(error: Error): ErrorBoundaryState {
-    return { errorMessage: error.toString() };
-  }
-
-  componentDidCatch(error: Error): void {
-    console.error(error);
-  }
-
-  render(): ReactNode {
-    if (this.state.errorMessage) {
-      return <p>{this.state.errorMessage}</p>;
-    }
-    return this.props.children;
-  }
-}
